Add Benchmark
This starts work on #7. Currently this provides a comparison between
IronRure and .NET regex. Initial results seem to suggest that
transcoding the haystack to UTF8 seems to dominate the benchmarks for
IronRure so overloads where the UTF8 could be provided directly were
added.

The text to search comes from Project Gutenberg.