Move to `SafeHandle` for Unmanaged Pointers (#38)
* Move to `SafeHandle` for Unmanaged Pointers

This is an API breaking change but should make the use of unsafe
poiters in this library more reliable in some corner cases. Safe
handles will ensure that the attached resources are always correctly
released, even in exceptional circumstances.

This also allows the `RureFfi` to be a tad more type safe in what it
accepts. It's bad that we were somehow _less type safe_ than C. :-/

* Bump Version

This new handle approach is not API or ABI backwards compatible. It is
however _most likely_ source compatible.